Move tlogserver main file dir from cmd to tlog/tlogserver.
