+++ tlog/tlogserver/server/end_to_end_test.go
@@ -1,4 +1,4 @@
-package tlogserver
+package server
 
 import (
 	"testing"
